[klibc] ipconfig: Write $DOMAINSEARCH as domain-search
This patch requests domain-search option to DHCP servers, and parses
domain-search option value in DHCP ACK packet when ipconfig uses DHCP.
This patch writes $DOMAINSEARCH as the value to /tmp/net-$DEVICE.conf file.

> ipconfig: ignores dhcp options
> http://bugs.debian.org/627166
This Debian Bug report requests the option value.

The format of $DOMAINSEARCH is domain name list which is separated with ' '.
For example, $DOMAINSEARCH is either 'foo1.bar. foo2.bar.', 'foo.', or ''.

Signed-off-by: KUMAAN <9maaan@gmail.com>
Acked-by: H. Peter Anvin <hpa@zytor.com>
Signed-off-by: maximilian attems <max@stro.at>